mappers

+++ src/main/kotlin/com/example/kotlinstudies/dto/PersonDto.kt
@@ -1,5 +1,5 @@
 package com.example.kotlinstudies.dto
 
-data class PersonDto (val name: String, val surname: String, val age: Long) {
+data class PersonDto (val name: String, val surname: String, val age: Int) {
 
 }
